# packages/backend/openapi.yaml
openapi: 3.1.0
info:
  title: sanposcape API
  description: 散歩支援アプリ sanposcape のバックエンド API
  version: 0.1.0
paths:
...
  /pins:
    post:
      tags:
      - pins
      summary: Create Pin
      description: 'ピンの記録を保存する。


        `client_pin_id` の再送は同じピンとして扱い、新規作成のみ 201、冪等な再送は 200 を

        返す（内容が違っても既存をそのまま返す。`walks` と同じ）。`sanpo_map_id` を省略すると

        自分の既定地図（無ければ「最初の地図」を同一トランザクションで作成）に登録される。'
      operationId: create_pin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PinCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PinRead'
        '401':
          description: Not authenticated
        '413':
          description: Request body too large
        '200':
          description: 'Idempotent replay: an existing pin with the same client_pin_id'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PinRead'
        '403':
          description: Permission denied
        '404':
          description: Sanpo map not found
        '409':
          description: Photo upload not ready, or storage quota exceeded (see body
            `code` to distinguish them)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PinConflictErrorRead'
        '422':
          description: Validation error
        '503':
          description: Photo storage unavailable
      security:
      - HTTPBearer: []
  /pins/{pin_id}/photos:
    post:
      tags:
      - pins
      summary: Add Pin Photos
      description: 'ピンに写真を追加する（1リクエストあたり1〜10枚。ピン全体の枚数は無制限, B-Y2）。


        既にこのピンに紐付いている `upload_id` の再送は成功扱い。別のピンに紐付け済みの

        `upload_id`・存在しない/他人の/期限切れの枠・デコード不可な画像は 409

        （`code: "photo_upload_not_ready"`）。確定時の実サイズで容量上限を超える場合も 409

        （`code: "storage_quota_exceeded"`）。他人のピン・存在しないピンは 404。'
      operationId: add_pin_photos
      security:
      - HTTPBearer: []
      parameters:
      - name: pin_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Pin Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PinPhotosAdd'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PinPhotoListRead'
        '401':
          description: Not authenticated
        '413':
          description: Request body too large
        '403':
          description: Permission denied
        '404':
          description: Pin not found
        '409':
          description: Photo upload not ready, or storage quota exceeded (see body
            `code` to distinguish them)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PinConflictErrorRead'
        '422':
          description: Validation error
        '503':
          description: Photo storage unavailable
  /pin-photo-uploads:
    post:
      tags:
      - pins
      summary: Create Pin Photo Upload
      description: '写真のアップロード枠（presigned POST）を発行する。


        応答の `upload.fields` を multipart フォームで `upload.url` へ POST する

        （フィールドの意味はクライアントが解釈しない）。上限は応答の `max_byte_size` が正。'
      operationId: create_pin_photo_upload
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PinPhotoUploadCreate'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PinPhotoUploadRead'
        '401':
          description: Not authenticated
        '409':
          description: 'Storage quota exceeded (body code: "storage_quota_exceeded")'
        '413':
          description: Request body too large, or photo exceeds the size limit
        '422':
          description: Validation error
        '429':
          description: Too many pending uploads
        '503':
          description: Photo storage unavailable
      security:
      - HTTPBearer: []
components:
  schemas:
...
    GeoPoint:
      properties:
        latitude:
          type: number
          maximum: 90.0
          minimum: -90.0
          title: Latitude
        longitude:
          type: number
          maximum: 180.0
          minimum: -180.0
          title: Longitude
      type: object
      required:
      - latitude
      - longitude
      title: GeoPoint
...
    PinConflictErrorRead:
      properties:
        detail:
          type: string
          title: Detail
        code:
          type: string
          enum:
          - storage_quota_exceeded
          - photo_upload_not_ready
          title: Code
      type: object
      required:
      - detail
      - code
      title: PinConflictErrorRead
      description: '`POST /pins`・`POST /pins/{pin_id}/photos` の 409 応答本体（PR #93 T15）。


        `detail` は既存クライアントとの後方互換のため固定文言のまま維持する。`code` は

        mobile が「容量超過」（`storage_quota_exceeded`）と「写真の準備ができていない」

        （`photo_upload_not_ready`。未完了・期限切れ・デコード不可などをまとめて表す,

        ADR-009 決定9）を区別するために新規追加した機械可読な値。'
    PinCreate:
      properties:
        client_pin_id:
          type: string
          format: uuid
          title: Client Pin Id
        sanpo_map_id:
          type: string
          format: uuid
          title: Sanpo Map Id
        name:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Name
        memo:
          anyOf:
          - type: string
            maxLength: 1000
          - type: 'null'
          title: Memo
        location:
          $ref: '#/components/schemas/GeoPoint'
        tags:
          items:
            type: string
            maxLength: 200
            minLength: 1
          type: array
          maxItems: 10
          title: Tags
        photo_upload_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 10
          title: Photo Upload Ids
        client_walk_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Client Walk Id
      type: object
      required:
      - client_pin_id
      - location
      title: PinCreate
    PinPhotoListRead:
      properties:
        items:
          items:
            $ref: '#/components/schemas/PinPhotoRead'
          type: array
          title: Items
        photo_count:
          type: integer
          title: Photo Count
      type: object
      required:
      - items
      - photo_count
      title: PinPhotoListRead
    PinPhotoRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        upload_id:
          type: string
          format: uuid
          title: Upload Id
        position:
          type: integer
          title: Position
        width:
          type: integer
          title: Width
        height:
          type: integer
          title: Height
        byte_size:
          type: integer
          title: Byte Size
        content_type:
          type: string
          title: Content Type
        thumbnail:
          anyOf:
          - $ref: '#/components/schemas/PinPhotoThumbnailRead'
          - type: 'null'
        urls_expire_at:
          type: string
          format: date-time
          title: Urls Expire At
        uploaded_by_user_id:
          type: string
          format: uuid
          title: Uploaded By User Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - upload_id
      - position
      - width
      - height
      - byte_size
      - content_type
      - thumbnail
      - urls_expire_at
      - uploaded_by_user_id
      - created_at
      title: PinPhotoRead
    PinPhotoThumbnailRead:
      properties:
        url:
          type: string
          title: Url
        width:
          type: integer
          title: Width
        height:
          type: integer
          title: Height
      type: object
      required:
      - url
      - width
      - height
      title: PinPhotoThumbnailRead
    PinPhotoUploadCreate:
      properties:
        content_type:
          type: string
          const: image/jpeg
          title: Content Type
        byte_size:
          type: integer
          minimum: 1.0
          title: Byte Size
      type: object
      required:
      - content_type
      - byte_size
      title: PinPhotoUploadCreate
      description: '`POST /pin-photo-uploads` のリクエスト。


        `width`/`height` は受け取らない（サーバーが確定時にデコードして実寸を得るため）。'
    PinPhotoUploadRead:
      properties:
        upload_id:
          type: string
          format: uuid
          title: Upload Id
        upload:
          $ref: '#/components/schemas/PresignedUploadRead'
        expires_at:
          type: string
          format: date-time
          title: Expires At
        max_byte_size:
          type: integer
          title: Max Byte Size
      type: object
      required:
      - upload_id
      - upload
      - expires_at
      - max_byte_size
      title: PinPhotoUploadRead
    PinPhotosAdd:
      properties:
        photo_upload_ids:
          items:
            type: string
            format: uuid
          type: array
          maxItems: 10
          minItems: 1
          title: Photo Upload Ids
      type: object
      required:
      - photo_upload_ids
      title: PinPhotosAdd
      description: '`POST /pins/{pin_id}/photos` のリクエスト。'
    PinRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        client_pin_id:
          type: string
          format: uuid
          title: Client Pin Id
        sanpo_map:
          $ref: '#/components/schemas/SanpoMapSummaryRead'
        name:
          anyOf:
          - type: string
          - type: 'null'
          title: Name
        memo:
          anyOf:
          - type: string
          - type: 'null'
          title: Memo
        location:
          $ref: '#/components/schemas/GeoPoint'
        tags:
          items:
            $ref: '#/components/schemas/PinTagRead'
          type: array
          title: Tags
        photos:
          items:
            $ref: '#/components/schemas/PinPhotoRead'
          type: array
          title: Photos
        photo_count:
          type: integer
          title: Photo Count
        created_by_user_id:
          type: string
          format: uuid
          title: Created By User Id
        client_walk_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Client Walk Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - client_pin_id
      - sanpo_map
      - name
      - memo
      - location
      - tags
      - photos
      - photo_count
      - created_by_user_id
      - client_walk_id
      - created_at
      - updated_at
      title: PinRead
    PinTagRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        label:
          type: string
          title: Label
        created_by_user_id:
          type: string
          format: uuid
          title: Created By User Id
      type: object
      required:
      - id
      - label
      - created_by_user_id
      title: PinTagRead
...
    PresignedUploadRead:
      properties:
        url:
          type: string
          title: Url
        fields:
          additionalProperties:
            type: string
          type: object
          title: Fields
      type: object
      required:
      - url
      - fields
      title: PresignedUploadRead
      description: presigned POST の応答。`fields` はクライアントが解釈せず、そのまま multipart で送る。
...
    SanpoMapSummaryRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        is_default:
          type: boolean
          title: Is Default
      type: object
      required:
      - id
      - name
      - is_default
      title: SanpoMapSummaryRead
      description: 他ドメイン（`pins/schemas.py` の `PinRead.sanpo_map`）が埋め込む要約表現。
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer